ZUGFeRD — generalise org-master window spec; run en_US + de_DE
Repurpose the einvoice-specific seller-tax-fields spec into a window-generic spec
for window 540676 'Organisation Stammdaten' (rename einvoice-seller-tax-fields.spec.js
-> organisation-stammdaten-window.spec.js). The displayed-field checks now read from an
EXPECTED_DISPLAYED_FIELDS extension point so future issues that make C_BPartner columns
visible in this window append to it rather than creating a new einvoice-named spec.

Language independence: assert only on the C_BPartner DB ColumnName (language-invariant,
via layout JSON field.field) — never a UI caption; drop the German cosmetic labels; and
run the test in BOTH en_US and de_DE to prove the login + layout flow is language-independent.
Update TEST-COVERAGE.md section 42 accordingly.
